less msvc warns

+++ core/buffers/kernels/spatialGradient.cpp
@@ -56,11 +56,11 @@ G12Buffer* SpatialGradient::findCornerPoints(double scaler, int apperture)
     /* TODO: Redo this to simplify for buffers */
     if (v == 5)
     {
-    BufferProcessor<SpatialGradient, SpatialGradient, Blur5Horisontal, ScalarAlgebraVector33> blurerH;
-    BufferProcessor<SpatialGradient, SpatialGradient, Blur5Vertical, ScalarAlgebraVector33>   blurerV;
+        BufferProcessor<SpatialGradient, SpatialGradient, Blur5Horisontal, ScalarAlgebraVector33> blurerH;
+        BufferProcessor<SpatialGradient, SpatialGradient, Blur5Vertical, ScalarAlgebraVector33>   blurerV;
 
-    blurerV.process(&start   , &blurVert);
-    blurerH.process(&blurVert, &blur);
+        blurerV.process(&start   , &blurVert);
+        blurerH.process(&blurVert, &blur);
     }
     else
     {
@@ -72,9 +72,9 @@ G12Buffer* SpatialGradient::findCornerPoints(double scaler, int apperture)
             {
                 Vector3dd sum = Vector3dd::Zero();
 
-                for (int dy = i - v / 2 ; dy <= i + v / 2; dy++)
+                for (int dy = i - v / 2 ; dy <= (int)i + v / 2; dy++)
                 {
-                    for (int dx = j - v / 2 ; dx <= j + v / 2; dx++)
+                    for (int dx = j - v / 2 ; dx <= (int)j + v / 2; dx++)
                     {
                         if (!start->isValidCoord(dy,dx))
                             continue;
